refactor: add ProposalActions folder
